rust-td-payload: split test code from sample code
The rust-td-payload is a sample TD module for TDX hardware platforms.
Split out tested related code from it into tests/test-td-payload crae.
