+++ rust-td-payload/Cargo.toml
@@ -1,43 +1,26 @@
 [package]
 name = "rust-td-payload"
 version = "0.1.0"
+description = "A sample TD module/driver for TDX hardware platforms"
+repository = "https://github.com/confidential-containers/td-shim"
+homepage = "https://github.com/confidential-containers"
+license = "BSD-2-Clause-Patent"
 edition = "2018"
 
-[features]
-cet-ss = []
-benches = ["benchmark"]
-# add build process
-# build = "build.rs"
-
-#[build-dependencies]
-#cc = { version = "1.0.62", default-features = false }
-
-
 [dependencies]
-spin = "0.9.2"
-r-efi = "3.2.0"
+benchmark = { path = "../benchmark", optional = true }
+chrono = { version = "0.4.19", default-features = false, features = ["serde"]}
 linked_list_allocator = "0.9.0"
 log = "0.4.13"
-r-uefi-pi =  { path = "../r-uefi-pi" }
-uefi-pi =  { path = "../uefi-pi" }
-tdx-tdcall = { path = "../tdx-tdcall" }
-tdx-logger =  { path = "../tdx-logger" }
-tdx-exception =  { path = "../tdx-exception" }
-td-layout = { path = "../td-layout" }
+r-efi = "3.2.0"
 serde = { version = "1.0", default-features = false, features = ["derive"]}
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
-chrono = { version = "0.4.19", default-features = false, features = ["serde"]}
-scroll = { version = "0.10", default-features = false, features = ["derive"]}
-x86 = "0.41.0"
-benchmark = { path = "../benchmark", optional = true }
-test-runner-client = { path = "../devtools/test-runner-client" }
+td-layout = { path = "../td-layout" }
+tdx-tdcall = { path = "../tdx-tdcall" }
+tdx-logger =  { path = "../tdx-logger" }
+uefi-pi =  { path = "../uefi-pi" }
 
-bootloader = "=0.10.9"
-
-
-[dependencies.lazy_static]
-version = "1.0"
-features = ["spin_no_std"]
-
-[package.metadata.bootloader]
-map-physical-memory = true
+[features]
+default = ["cet-ss"]
+cet-ss = []
+benches = ["benchmark"]
