+++ rust-td-payload/src/lib.rs
@@ -1,0 +1,31 @@
+// Copyright © 2019 Intel Corporation
+// Copyright (c) 2022 Alibaba Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#![no_std]
+#![feature(alloc_error_handler)]
+
+#[macro_use]
+extern crate alloc;
+
+#[cfg(not(test))]
+#[alloc_error_handler]
+#[allow(clippy::empty_loop)]
+fn alloc_error(_info: core::alloc::Layout) -> ! {
+    log::info!("alloc_error ... {:?}\n", _info);
+    panic!("deadloop");
+}
+
+mod json;
+pub use json::json_test;
